Added examples for a BQ command, and Image downloader
The BQ Command is two fold, one for persisting JSON data to a table,
while the other is capable of reading rows and converting them into JSON
for the context to consume.

The image download command reads images from the context and downloads
them to a GCS storage bucket. If the bucket does not exist, the commanad
attempts to create it. No images will be downloaded if the storage
bucket does not exist.